# api/openapi/running-coach.yaml
openapi: 3.1.1
info:
  title: Running Coach API
  version: 0.0.1
  description: Contrato público de Running Coach.
  contact:
    name: Running Coach
servers:
  - url: /api
    description: API bajo el mismo origen.
tags:
  - name: Access
    description: Ciclo de vida de la sesión actual.
paths:
  /sessions:
    post:
      operationId: createSession
      tags: [Access]
      summary: Crear una sesión
      description: >-
        Autentica una cuenta activa mediante correo y contraseña. Credenciales, estado o
        cuenta inválidos producen la misma respuesta pública y no permiten enumeración.
      security:
        - csrfCookie: []
          csrfToken: []
      requestBody:
        $ref: '#/components/requestBodies/SessionCreationRequest'
      responses:
        '201':
          $ref: '#/components/responses/SessionCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'
        '401':
          $ref: '#/components/responses/SessionCreationRejectedProblem'
        '403':
          $ref: '#/components/responses/CsrfValidationFailedProblem'
        '429':
          $ref: '#/components/responses/RateLimitExceededProblem'
        '500':
          $ref: '#/components/responses/InternalServerErrorProblem'
  /sessions/current:
    get:
      operationId: getCurrentSession
      tags: [Access]
      summary: Consultar la sesión actual
      description: >-
        Resuelve la cuenta desde la cookie opaca y devuelve únicamente su identidad,
        rol inmutable y estado vigente.
      security:
        - opaqueSession: []
      responses:
        '200':
          $ref: '#/components/responses/CurrentSessionResponse'
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '500':
          $ref: '#/components/responses/InternalServerErrorProblem'
    delete:
      operationId: deleteCurrentSession
      tags: [Access]
      summary: Revocar la sesión actual
      description: >-
        Revoca la sesión asociada a la cookie, elimina la cookie de sesión y rota el
        contexto CSRF. Repetir la operación no restaura ni duplica efectos.
      security:
        - opaqueSession: []
          csrfCookie: []
          csrfToken: []
      responses:
        '204':
          $ref: '#/components/responses/SessionRevokedResponse'
        '401':
          $ref: '#/components/responses/AuthenticationRequiredProblem'
        '403':
          $ref: '#/components/responses/CsrfValidationFailedProblem'
        '500':
          $ref: '#/components/responses/InternalServerErrorProblem'
components:
  securitySchemes:
    opaqueSession:
      type: apiKey
      in: cookie
      name: __Host-pmv_session
      description: >-
        Sesión opaca emitida por el backend en una cookie Secure, HttpOnly,
        SameSite=Lax, Path=/ y sin Domain.
    csrfCookie:
      type: apiKey
      in: cookie
      name: __Host-pmv_csrf
      description: >-
        Valor CSRF legible por la aplicación cliente, emitido con Secure, SameSite=Lax, Path=/ y sin
        Domain. Debe coincidir con X-CSRF-TOKEN.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: >-
        Token CSRF exigido en solicitudes que modifican estado. Debe coincidir con la
        cookie __Host-pmv_csrf y nunca se transporta en la URL.
  schemas:
    SessionCreation:
      type: object
      description: Credenciales presentadas para crear una sesión opaca.
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: Correo de acceso; se normaliza y compara de forma canónica.
        password:
          type: string
          minLength: 12
          maxLength: 128
          description: Contraseña enviada únicamente en el cuerpo HTTPS protegido.
    CurrentSession:
      type: object
      description: Identidad y atributos vigentes resueltos desde la sesión opaca.
      additionalProperties: false
      required: [accountId, role, accountStatus]
      properties:
        accountId:
          type: string
          format: uuid
          description: Identificador opaco de la cuenta autenticada.
        role:
          $ref: '#/components/schemas/AccountRole'
        accountStatus:
          $ref: '#/components/schemas/AccountStatus'
    AccountRole:
      type: string
      description: Rol único e inmutable de la cuenta.
      enum: [administrador, entrenador, corredor]
    AccountStatus:
      type: string
      description: Estado vigente de la cuenta consultado para cada operación protegida.
      enum: [pending_activation, active, disabled, pending_reactivation, cancelled]
    Problem:
      type: object
      description: Error conforme a RFC 9457, sin secretos ni datos personales innecesarios.
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Código estable de aplicación.
    Cursor:
      type: string
      minLength: 1
      description: Cursor opaco ligado al orden y filtros de una colección.
    PageLimit:
      type: integer
      minimum: 1
      maximum: 100
      default: 50
      description: Número máximo acotado de elementos solicitados.
    LocalDate:
      type: string
      format: date
    OffsetDateTime:
      type: string
      format: date-time
  parameters:
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Cursor'
    limit:
      name: limit
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/PageLimit'
  requestBodies:
    SessionCreationRequest:
      description: Credenciales para crear una sesión opaca.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SessionCreation'
          examples:
            synthetic:
              value:
                email: runner.synthetic@example.invalid
                password: <synthetic-password>
  headers:
    CurrentSessionLocation:
      description: URI del recurso contextual que representa la sesión recién creada.
      schema:
        type: string
        format: uri-reference
        const: /api/sessions/current
    SetCookie:
      description: >-
        Uno o varios campos Set-Cookie. La creación de sesión emite __Host-pmv_session y
        rota __Host-pmv_csrf; el cierre elimina la primera y rota la segunda.
      schema:
        type: array
        items:
          type: string
  examples:
    SyntheticRunnerSession:
      summary: Sesión sintética de corredor activo.
      value:
        accountId: 00000000-0000-4000-8000-000000000001
        role: corredor
        accountStatus: active
  responses:
    SessionCreatedResponse:
      description: Sesión creada y cookies de sesión y CSRF rotadas.
      headers:
        Location:
          $ref: '#/components/headers/CurrentSessionLocation'
        Set-Cookie:
          $ref: '#/components/headers/SetCookie'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CurrentSession'
          examples:
            syntheticRunner:
              $ref: '#/components/examples/SyntheticRunnerSession'
    CurrentSessionResponse:
      description: Representación mínima de la sesión actual.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CurrentSession'
          examples:
            syntheticRunner:
              $ref: '#/components/examples/SyntheticRunnerSession'
    SessionRevokedResponse:
      description: Sesión revocada; la respuesta no contiene cuerpo.
      headers:
        Set-Cookie:
          $ref: '#/components/headers/SetCookie'
    BadRequestProblem:
      description: Solicitud inválida.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            invalidRequest:
              value:
                type: /problems/invalid-request
                title: La solicitud no es válida
                status: 400
                code: invalid_request
    AuthenticationRequiredProblem:
      description: La sesión no existe, ha caducado o ha sido revocada.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            authenticationRequired:
              value:
                type: /problems/authentication-required
                title: Es necesario iniciar sesión
                status: 401
                code: authentication_required
    SessionCreationRejectedProblem:
      description: >-
        No se crea una sesión. La respuesta no diferencia cuenta, contraseña o estado
        inválidos.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            rejected:
              value:
                type: /problems/session-creation-rejected
                title: No se ha podido iniciar sesión
                status: 401
                code: session_creation_rejected
    ForbiddenProblem:
      description: El actor no puede realizar la operación.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    CsrfValidationFailedProblem:
      description: El contexto CSRF o el origen de la solicitud no son válidos.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            rejected:
              value:
                type: /problems/csrf-validation-failed
                title: La solicitud no ha superado la validación de seguridad
                status: 403
                code: csrf_validation_failed
    NotFoundProblem:
      description: El recurso no existe o no es visible para el actor.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimitExceededProblem:
      description: >-
        Se ha superado un límite de acceso. La respuesta no identifica el contador
        alcanzado ni revela si la cuenta existe.
      headers:
        Retry-After:
          description: Segundos mínimos antes de intentar otra vez.
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            limited:
              value:
                type: /problems/rate-limit-exceeded
                title: No se puede intentar el acceso en este momento
                status: 429
                code: rate_limit_exceeded
    InternalServerErrorProblem:
      description: Error interno sin información sensible.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            internalError:
              value:
                type: /problems/internal-error
                title: Se ha producido un error interno
                status: 500
                code: internal_error